requirements.txt
sympy
feynman
fractions
itertools
scipy
IPython.display
matplotlib.pyplot
numpy